refactor Titre and font
