+++ app/src/main/java/com/github/se/polyfit/ui/compose/Font.kt
@@ -1,0 +1,13 @@
+package com.github.se.polyfit.ui.compose
+
+import androidx.compose.ui.text.font.Font
+import androidx.compose.ui.text.font.FontFamily
+import androidx.compose.ui.text.font.FontWeight
+import com.github.se.polyfit.R
+
+val kaiseiFont = FontFamily(
+    Font(R.font.kaiseitokumin_bold, FontWeight.Bold),
+    Font(R.font.kaiseitokumin_regular, FontWeight.Normal),
+    Font(R.font.kaiseitokumin_extrabold, FontWeight.ExtraBold),
+    Font(R.font.kaiseitokumin_medium, FontWeight.Medium)
+)
